CONVERT: contacts.js to contacts.ts file
